backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-multipart
bcrypt
passlib
passlib[bcrypt]
python-jose[cryptography]
pycryptodome
pycryptodomex
cryptography
python-slugify